requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
mysqlclient==2.1.1
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2